requirements.txt
aiohttp==3.8.5
botorch==0.8.5
ConfigSpace==0.7.1
emcee==3.1.4
gpytorch==1.10
h5py==3.9.0
HEBO==0.3.5
jax==0.2.24
langchain==0.1.9
nashpobench2api==1.0.0
openai==0.28.0
openml==0.14.2
optuna==3.3.0
pandas==1.3.5
parzen_estimator==0.5.8
pyarrow==12.0.1
scikit_learn==1.0.2
scikit_optimize==0.9.0
scipy==1.12.0
smac==1.4.0
tiktoken==0.6.0
uncertainty_toolbox==0.1.1
pynisher==0.6.4
numpy==1.23.4
transformers 
accelerate 
wandb